integrated score & leaderboard into game
