+++ client/style.css
@@ -27,14 +27,14 @@ body {
 #leader-board {
   border-style: solid;
   border-radius: 25px;
-  width: 500px;
-  height: 475px;
+  width: 250px;
+  height: 300px;
   display: flex;
   flex-direction: column;
   align-items: center;
   position: absolute;
-  top: 75px; 
-  left: 100px; 
+  top: 165px; 
+  left: 125px; 
 }
 
 .go-to-lb {
@@ -65,8 +65,8 @@ body {
 
 .leader {
   border-bottom: solid;
-  height: 75px;
-  width: 300px;
+  height: 45px;
+  width: 175px;
   text-align: center;
   padding-top: 25px
 }
